initial inventory model definitions complete
